kariim

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -11,7 +11,7 @@
 int get_bit(unsigned long int n, unsigned int index)
 {
 	if (index >= sizeof(n) * 8)
-		return (-1);
+		return ((unsigned int)-1);
 
-	return (n >> index & 1);
+	return (int)(n >> index & 1);
 }
